The camera switches well now
I am using a RemoteTransform2D node and life got very good. I can just link and unlink it
to the Camera to make it follow the character. I would just like to stop the character
from moving during the transition between static and dynamic camera, but this is proving
to be harder than I thought due to how I designed the timeCanMove system. For now I added a new
variable named "cutscene_is_running", but I'm not completely satisfied with this solution. Maybe I'll get
back to it in the future (or will I?).