Jump successfully implemented!
Now Time.checkIfMoving() simply takes the input and then, in case of movement, emits is_moving(). Then the World makes all the checks to make sure the movement is legal. Finally, the World itself calls Time.move() with the received coordinates, to make the movement actually happen. The variable is_a_jump ensures not to lose sand twice in case of a movement+jump situation.